The user can still click the 'Estimate Sigma X' button if they want but they can now tweak any value that they want and NOT have it affect the Smoothness or Sigma X values.
While manually entering values for Y Start or Y End if the values are 'incorrect' , ie, the Y Start is greater than the Y End or vice versa then the Reconstruction Area box does NOT update and the Y Start and Y End input text fields put a RED outline around themselves to indicate an issue and the Reconstruction Button disables itself. After the user corrects the issue then the Reconstruction Button becomes available and the red outline is removed. I think this will work pretty good to make sure the reconstruction area is sane. There are also checks after the user clicks the Reconstruction button to make sure the geometry is correct before proceeding. I also display a message in the bottom of the window in the status bar area telling the user what is wrong.

 Any loaded "Config" file is added to the list of Recent files. This way when TomoGui is launched the user can select a configuration file from the list of Recent files and it will load up the mrc or rec file and set all the parameters (I think: Sigma X seems not correct but you can check that). I know this adds one more step but this way should ensure that the values are appropriate for a reconstruction.